+++ SmartComplexSolution/ThanalSoft.SmartComplex.DataObjects/Common/State.cs
@@ -28,6 +28,6 @@ namespace ThanalSoft.SmartComplex.DataObjects.Common
 
         [DataMember]
         [NotMapped]
-        public override int LastUpdatedBy { get; set; }
+        public override Int64 LastUpdatedBy { get; set; }
     }
 }
