+++ SmartComplexSolution/ThanalSoft.SmartComplex.DataObjects/Complex/AmenityType.cs
@@ -26,7 +26,7 @@ namespace ThanalSoft.SmartComplex.DataObjects.Complex
 
         [DataMember]
         [NotMapped]
-        public override int LastUpdatedBy { get; set; }
+        public override Int64 LastUpdatedBy { get; set; }
 
         public virtual ICollection<ApartmentAmenity> ApartmentAmenities { get; set; }
